manual_file/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
pandas==2.2.2
openpyxl==3.1.5
pyarrow==17.0.0
unidecode==1.3.8
fastparquet